put variables in markdown files, and use that for release notes

+++ devscripts/generateReleaseNotes.js
@@ -1,20 +1,26 @@
-// designated for linux
+const fs = require(`fs`);
+const child_process = require(`child_process`);
+
+const daysOfTheWeek = [`Sun`, `Mon`, `Tue`, `Wed`, `Thu`, `Fri`, `Sat`];
+
+const months = [`Jan`, `Feb`, `Mar`, `Apr`, `May`, `Jun`, `Jul`, `Aug`, `Sep`, `Oct`, `Nov`, `Dec`]
 
 const args = process.argv.slice(2);
 
 const type = args.shift();
 
-if(type == `nightly`) {
-    const fs = require(`fs`);
-    const child_process = require(`child_process`);
-    
-    const commits = child_process.execSync(`git log $(git rev-list -n 1 $(git describe --tags --abbrev=0))...$(git rev-parse HEAD)`, {shell: `/bin/bash`}).toString().trim();
-    
-    const daysOfTheWeek = [`Sun`, `Mon`, `Tue`, `Wed`, `Thu`, `Fri`, `Sat`];
-    
-    const months = [`Jan`, `Feb`, `Mar`, `Apr`, `May`, `Jun`, `Jul`, `Aug`, `Sep`, `Oct`, `Nov`, `Dec`]
-    
-    const str = commits.slice(7).split(`\n\ncommit `).map(s => {
+const tagList = child_process.execSync(`git tag --sort=-creatordate`).toString().trim().split(`\n`).map(s => s.trim());
+const latestReleaseTag = tagList[0];
+const latestStableTag = tagList.filter(s => !s.includes(`-nightly`))[0];
+const previousTagCommit = child_process.execSync(`git rev-list -n 1 ${tagList.filter(s => !s.includes(`-nightly`))[0]}`).toString().trim();
+const currentCommit = child_process.execSync(`git rev-parse HEAD`).toString().trim();
+
+console.log(`latest release tag: ${latestReleaseTag} -- latest stable tag: ${latestStableTag}`)
+
+const variables = {
+    version: require(`../package.json`).version,
+    releaseTitle: process.env["RELEASE_TITLE"] || null,
+    commitList: child_process.execSync(`git log ${previousTagCommit}...${currentCommit}`).toString().trim().slice(7).split(`\n\ncommit `).map(s => {
         const hash = s.split(`\n`)[0];
         const hashLink = `[**${hash.slice(0, 7)}**](https://github.com/ezytdl/ezytdl/commit/${hash})`;
         //console.log(hashLink);
@@ -31,15 +37,19 @@ if(type == `nightly`) {
         const parsed = `> - ${author}: ${hashLink} / ${dateStr}\n> \n> ${s.split(`\n`).slice(4).map(s => s.trim()).join(`\n> `)}\n`;
     
         return parsed;
-    });
-    
-    console.log(str);
+    }).join(`\n`)
+};
 
-    let processed = fs.readFileSync(`./res/releaseNotes/nightly.md`).toString() + str.join(`\n`)
+if(fs.existsSync(`./res/releaseNotes/${type}.md`)) {
+    const variableRegex = /{([^}]+)}/g;
+
+    let processed = fs.readFileSync(`./res/releaseNotes/${type}.md`).toString().replace(variableRegex, (match, key) => variables[key] || match);
+
+    if(processed.match(variableRegex)) throw new Error(`Failed to replace all variables! Missing:` + processed.match(variableRegex).map(s => `\n- Missing ${s.slice(1, -1)} -- got ${variables[s.slice(1, -1)]}`).join(``))
 
     if(args.length > 0) {
         processed = processed.split(`--------------`)[0] + `--------------\n\n### Release Notes\n\n${args.join(` `)}\n\n--------------\n\n` + processed.split(`--------------`)[1]
     }
-    
+
     fs.writeFileSync(`./release-notes.md`, processed)
-}
+} else console.error(`No release notes template found for type "${type}"!`)
